+++ utils/buffer_builder.cpp
@@ -1,5 +1,5 @@
 #include "buffer_builder.h"
-buffer_builder::buffer_builder(int8_t* buffer, int size) :
+buffer_builder::buffer_builder(int8_t* buffer, size_t size) :
 	buffer(buffer), pointer(buffer), size(size)
 {
 }
